+++ CHANGELOG.md
@@ -5,8 +5,10 @@
 Unreleased
 
 - Added the `--join` option to display the data related to `keypoint show`, `analysis show` and `drawing show` commands at the same time.
-- Removed setting API URL in `configure` command.
+- Added `--bg-rule` option in `draw` command.
+- Made it possible to draw background using `--rule-file`.
 - Improved the display of JSON data response and fixed to not display binary data response when using `-v` option.
+- Removed setting API URL in `configure` command.
 - Fixed a bug that prevented download without keypoint id.
 
 ## 1.0.1